% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storr_bdb.R
\name{storr_bdb}
\alias{driver_bdb}
\alias{storr_bdb}
\title{Berkeley DB storr driver}
\usage{
storr_bdb(path, type = "BTREE", default_namespace = "objects")

driver_bdb(path, type = "BTREE")
}
\arguments{
\item{path}{Path for the database.  If \code{NULL} a transient
in-memory database will be created.}

\item{type}{Storage type for all the data.  Sensible options are
\code{BTREE} and \code{HASH}.}

\item{default_namespace}{Default namespace (see
\code{\link{storr}}).}
}
\description{
Berkeley DB storr driver.
}
\section{Design considerations}{


Because there is no easy way of iterating over all keys (without
retrieving all data), the data is stored in two different
databases within the file; one called \code{data} the other called
\code{keys}.  This should make \code{list_keys} a bit faster, but
\code{list_data} will always be very slow and memory intensive.
}
\examples{
st <- storr_bdb(NULL)
}

